Convert Firefox History to Sanitised SQLite DB
Add files to create a SQLite database using the Firefox browsing history
as a base. Browsing history is stripped of domains that match a user
defined, comma delineated list of domains called .pearsignore located
in the root of the home directory.

Output to stdout through print statements still enabled to show the
response code of the url, which urls are being omitted from collection
and which  urls are being committed to the database. Only urls that
begin with 'http' and return a response code of 200 are recorded. This
is useful to assist in creating your .pearsignore file.

Disk output will place a file called 'history.db' in the directory above
PeARS. This SQLite database consists of a table called 'History' with
four columns named: 'Id', 'URL', 'TITLE', and 'BODY'.

BeautifulSoup and requests libraries are new requirements.